pyproject.toml
[tool.poetry]
name = "vestibulare"
version = "0.1.0"
description = ""
authors = ["Victor Hermon <victorsilva.silva6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pymssql = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
